backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
diffusers
transformers
accelerate
opencv-python
pillow
numpy
scipy
rembg
onnxruntime
